Adjust match confidence based on detected camera and increase for Vesper.
